LayoutTests/plugins/bindings-test-expected.txt

ALERT: typeof(objCPlugin)=object
ALERT: objCPlugin()=TypeError
ALERT: typeof(objCPluginFunction)=function
ALERT: objCPluginFunction()=test
ALERT: hello = hello (same)
ALERT: string = string (same)
ALERT: 42 = 42 (same)
ALERT: number = number (same)
ALERT: null = null (same)
ALERT: object = object (same)
ALERT: one,two = one,two (same)
ALERT: object = object (same)
ALERT: typeof(cPlugin)=function
ALERT: cPlugin()=1
ALERT: typeof(cPlugin)=object
ALERT: cPlugin()=TypeError
This page tests invoking as functions JavaScript objects that are provided by plugins.
